PowerInputs now take power from PowerOutputs
The previous system allowed a single input to have multiple sources. This meant that a door could have two switches connected to it. The switch that the door responds to was defined by the order the components are iterated. This new system prevents that from being possible. The new system also allows a single PowerOutput have multiple destinations, meaning that a single switch can control two doors.

The new system is better