Remove screenshot functionality
The tiny file dialogs doesn't actually work in this particular application. I've posted a stack overflow question about this. But it isn't necessary anyway. On a mac you can use Grab to take a screenshot of a window. You could also press CMD+SHIFT+3 and then crop the image. I'm sure it's equally simple on windows. This means that a screenshot key really isn't necessary.