Create Element base class
I really should stop wasting time on the UI and making some levels. Making levels for this game is much more difficult than I thought it would be.

I think my games might be getting too big for me to handle. In a game company, there are people for each aspect of the game. There are...
  -  designers
  -  animators
  -  artists
  -  programmers
  -  testers
  -  advertisers

I have to do all of those things. I can write code. I can (sort of) animate. I can test. That other stuff is not easy. I really want to be one of the people that write the code, nothing else.