Create level 6
I've renamed the old level 6 to level 10 because after I finished the paper version I realised that it's a bit to tricky for level 6.

A preprocessor for JSON would be really useful. Level 6 is 1388 lines copy-and-pasted JSON. A preprocessor would be able to do the copy-and-pasting automatically. A preprocessor designed specifically for JSON would definitely minify the output (remove whitespace). This would make loading levels a little bit faster.