Start converting levels to Jsonnet
Jsonnet is a language that compiles into JSON. It's made by google and is capable of make my level files really tiny. The game no longer needs to know about entity files or physics body files. The game only loads levels. Everything is stored in the level.

This is gonna be great when I finish it