Use an enum to rotate doors
The level creator will specify the cell of the base of the door and then specify the orientation. The position of the base does not need to be changed each time the orientation is changed. The next step is to move the origin of the local coordinate systems of all of the entities to the bottom left corner. This way, they can all use the same math to calculate orientation offsets.l