Animate locks
I did some math a found out that if the sprites were 20x20, they would look better. The math was right, these new lock sprites are 20x20 and they look a lot better than the old locks. 20x20 sprites will look good on 16:9 resolutions that are a multiple of 360. So the sprites will look good at the following resolutions

  -  640x360		scaled x1
  -  1280x720		scaled x2
  -  1920x1080		scaled x3
  -  2560x1440		scaled x4
  -  3200x1800		scaled x5
  -  3840x2160		scaled x6

20x20 sprites will look good at all of the common resolutions