Scale box, button and door to 20x20
I see why sprites are traditionally done at powers of two. If I want something that is 1/4 wide and in the center then I can do that at any power of two that is greater than 4. But this doesn't work with 20. I don't think anyone will care about this anyway