Create text rendering system
I was originally going to render the text onto a frame buffer and then render the frame buffer onto the scene but I quickly realised how complicated that was going to be. I decided to render the glyphs straight onto the scene. There isn't much text in this game so there isn't really any need to optimize text rendering.

I'm going to make the title screen as simple as possible. It will just be something like this:

+--------------------------------+
|                                        |
|           BUTTONS            |
|                                        |
|   Press any key to start   |
|                                        |
+--------------------------------+

I'll draw the whole thing onto a single image and the title screen will just be a single texture on a quad. I might make the "Press any key to start" text flash or something with the stencil buffer. I don't want to waste time making a complex and dynamic UI, I want to focus on making the game fun. That's all that matters.