Complete the title screen
The title screen actually leverages the game rendering system for its animation. The title screen creates an entity and renders it with the RenderingSystem. I think this shows how flexible the game engine is. The title screen only needs to describe the position and animation of the entity and the RenderingSystem only needs to read the position and animation data. There's no coupling to switches or boxes because those things only exist in Jsonnet. The game has no knowledge of anything other than components.